# docs/pmm_companion_openapi.yaml
openapi: 3.0.3
info:
  title: PMM Companion API
  description: Read-only API for PMM Companion UI with events, metrics, reflections, commitments, and live streaming.
  version: "1.0.0"

servers:
  - url: http://localhost:8001
    description: Local development server

paths:
  /events:
    get:
      summary: Get paginated events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: after_id
          in: query
          schema:
            type: integer
        - name: kind
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Events list
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  events:
                    type: array
                    items:
                      type: object
                  pagination:
                    type: object

  /metrics:
    get:
      summary: Get current metrics (IAS, GAS, OCEAN traits, stage)
      responses:
        '200':
          description: Current metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  metrics:
                    type: object
                    properties:
                      ias:
                        type: number
                      gas:
                        type: number
                      traits:
                        type: object
                      stage:
                        type: object

  /reflections:
    get:
      summary: Get reflections
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Reflections list

  /commitments:
    get:
      summary: Get commitments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: ["open", "closed", "all"]
      responses:
        '200':
          description: Commitments list

  /stream:
    get:
      summary: WebSocket for live updates
      responses:
        '101':
          description: WebSocket connection